chore: Move from S3 -> File module
