+++ api.planx.uk/modules/file/service/utils.ts
@@ -1,5 +1,5 @@
 import S3 from "aws-sdk/clients/s3";
-import { isLiveEnv } from "../helpers";
+import { isLiveEnv } from "../../../helpers";
 
 export function s3Factory() {
   return new S3({
