--- api.planx.uk/s3/index.ts
@@ -1,13 +1,0 @@
-import {
-  privateUploadController,
-  publicUploadController,
-  publicDownloadController,
-  privateDownloadController,
-} from "./controller";
-
-export {
-  privateUploadController,
-  publicUploadController,
-  publicDownloadController,
-  privateDownloadController,
-};
